requirements.txt
Flask==2.3.2
FastAPI==0.104.1
PyQt6==6.6.1
requests==2.31.0
Pydantic==2.5.0
SQLAlchemy==2.0.23
cryptography==41.0.7
loguru==0.7.2
python-dotenv==1.0.0
 PyJWT==2.10.1
email-validator==2.1.0
bleach==6.1.0
uvicorn==0.24.0
httpx==0.25.2
aiohttp==3.9.1
alembic==1.13.0
psutil==5.9.6
prometheus-client==0.19.0
beautifulsoup4==4.12.2
lxml==4.9.3